Move calculate_ells_for_interpolation to generators
